Refactored converter, removed unnessesary class ConversionState
